build: Update dev Python version to 3.10.2
And completely refactor code base by moving code into the `src/` and
tests into the `tests/` directories.

On top of that bump dependencies to latest versions, as well as
pre-commit hooks. Finally, install mypy as dev dependency and configure
it via `pyproject.toml`.
