+++ .github/workflows/ci.yml
@@ -1,4 +1,8 @@
-name: "ci"
+defaults:
+  run:
+    shell: "bash"
+
+name: "CI"
 
 on:
   push:
@@ -8,10 +12,16 @@ on:
     branches: ["master"]
 
 env:
-  COVERALLS_VERSION: "2.1.1"
-  DEV_PYTHON_VERSION: "3.8"
-  POETRY_VERSION: "1.0.10"
-  TWINE_VERSION: "3.1.1"
+  DEV_PYTHON_VERSION: "3.10.2"
+  # To check latest versions run,
+  #
+  # ```bash
+  # pip-latest-release coveralls poetry twine
+  # ```
+  #
+  COVERALLS_VERSION: "3.3.1"
+  POETRY_VERSION: "1.1.13"
+  TWINE_VERSION: "3.8.0"
 
 jobs:
   dev:
@@ -19,7 +29,7 @@ jobs:
 
     strategy:
       matrix:
-        os: ["ubuntu-latest", "macos-latest"]
+        os: ["ubuntu-latest", "macos-latest", "windows-latest"]
 
     runs-on: "${{ matrix.os }}"
 
@@ -32,7 +42,7 @@ jobs:
           python-version: "${{ env.DEV_PYTHON_VERSION }}"
 
       - name: "Install poetry"
-        uses: "dschep/install-poetry-action@v1.3"
+        uses: "snok/install-poetry@v1.3.0"
         with:
           version: "${{ env.POETRY_VERSION }}"
 
@@ -40,14 +50,14 @@ jobs:
         run: "poetry install --no-dev"
 
       - name: "Import package"
-        run: "poetry run python -c 'import pre_commit_run_hook_entry'"
+        run: "poetry run python3 -c 'from pre_commit_run_hook_entry.cli import main'"
 
   test:
     name: "Python ${{ matrix.python-version }}"
 
     strategy:
       matrix:
-        python-version: ["3.7", "3.8"]
+        python-version: ["3.7", "3.8", "3.9", "3.10"]
 
     runs-on: "ubuntu-latest"
 
@@ -60,34 +70,46 @@ jobs:
           python-version: "${{ matrix.python-version }}"
 
       - name: "Install poetry"
-        uses: "dschep/install-poetry-action@v1.3"
+        uses: "snok/install-poetry@v1.3.0"
         with:
           version: "${{ env.POETRY_VERSION }}"
 
+      - name: "Cache venv"
+        uses: "actions/cache@v2.1.7"
+        with:
+          path: ".venv"
+          key: "venv-${{ matrix.python-version }}-${{ hashFiles('poetry.lock') }}"
+
       - name: "Install package"
-        run: |
-          poetry install
+        run: "poetry install"
+
+      - name: "Cache mypy"
+        if: "${{ startsWith(env.DEV_PYTHON_VERSION, matrix.python-version) }}"
+        uses: "actions/cache@v2.1.7"
+        with:
+          path: ".mypy_cache"
+          key: "mypy-${{ matrix.python-version }}"
 
       - name: "Run pre-commit"
-        uses: "pre-commit/action@v2.0.0"
+        if: "${{ startsWith(env.DEV_PYTHON_VERSION, matrix.python-version) }}"
+        uses: "pre-commit/action@v2.0.3"
+        with:
+          token: "${{ secrets.GITHUB_TOKEN }}"
 
       - name: "Run tests"
-        run: |
-          set -e
-
-          if [ "${{ matrix.python-version }}" != "${{ env.DEV_PYTHON_VERSION }}" ]; then
-            export PYTEST_ADDOPTS="--no-cov"
-          fi
-
-          poetry run python -m pytest
+        run: "poetry run python3 -m pytest"
+        env:
+          STAGE: "test"
 
       - name: "Send report to coveralls"
-        if: "matrix.python-version == env.DEV_PYTHON_VERSION"
+        if: "${{ startsWith(env.DEV_PYTHON_VERSION, matrix.python-version) }}"
+        run: |
+          set -euo pipefail
+
+          python3 -m pip install coveralls==${{ env.COVERALLS_VERSION }}
+          python3 -m coveralls --source=github
         env:
           GITHUB_TOKEN: "${{ secrets.GITHUB_TOKEN }}"
-        run: |
-          python -m pip install coveralls==${{ env.COVERALLS_VERSION }}
-          python -m coveralls
 
   package:
     needs: "test"
@@ -104,7 +126,7 @@ jobs:
           python-version: "${{ env.DEV_PYTHON_VERSION }}"
 
       - name: "Install poetry"
-        uses: "dschep/install-poetry-action@v1.3"
+        uses: "snok/install-poetry@v1.3.0"
         with:
           version: "${{ env.POETRY_VERSION }}"
 
@@ -113,18 +135,20 @@ jobs:
 
       - name: "Check package"
         run: |
+          set -euo pipefail
+
           python -m pip install twine==${{ env.TWINE_VERSION }}
           python -m twine check dist/*
 
       - name: "Publish package"
-        if: "github.event_name == 'push' && startsWith(github.event.ref, 'refs/tags')"
+        if: "${{ github.event_name == 'push' && startsWith(github.event.ref, 'refs/tags/') }}"
         uses: "pypa/gh-action-pypi-publish@v1.3.1"
         with:
           user: "${{ secrets.PYPI_USERNAME }}"
           password: "${{ secrets.PYPI_PASSWORD }}"
 
   release:
-    if: "startsWith(github.ref, 'refs/tags')"
+    if: "${{ startsWith(github.ref, 'refs/tags/') }}"
     needs: "package"
     runs-on: "ubuntu-latest"
