+++ .pre-commit-config.yaml
@@ -1,29 +1,29 @@
 minimum_pre_commit_version: "1.17.0"
 repos:
   - repo: "https://github.com/commitizen-tools/commitizen"
-    rev: "v2.1.0"
+    rev: "v2.20.5"
     hooks:
       - id: "commitizen"
         stages: ["commit-msg"]
 
   - repo: "https://github.com/psf/black"
-    rev: "20.8b1"
+    rev: "22.1.0"
     hooks:
       - id: "black"
         name: "Format code (black)"
         exclude: ^docs/.*$
 
   - repo: "https://github.com/asottile/blacken-docs"
-    rev: "v1.8.0"
+    rev: "v1.12.1"
     hooks:
       - id: "blacken-docs"
         name: "Format docs (blacken-docs)"
         args: ["-l", "64"]
         additional_dependencies:
-          - "black==20.8b1"
+          - "black==22.1.0"
 
   - repo: "https://github.com/pre-commit/pre-commit-hooks"
-    rev: "v3.2.0"
+    rev: "v4.1.0"
     hooks:
       - id: "end-of-file-fixer"
       - id: "trailing-whitespace"
@@ -31,40 +31,42 @@ repos:
       - id: "check-toml"
       - id: "check-yaml"
 
-  - repo: "https://gitlab.com/PyCQA/flake8"
-    rev: "3.8.3"
+  - repo: "https://github.com/PyCQA/flake8"
+    rev: "4.0.1"
     hooks:
       - id: "flake8"
         name: "Lint code (flake8)"
         additional_dependencies: &flake8_additional_dependencies
-          - "flake8==3.8.3"
-          - "flake8-broken-line==0.2.1"
-          - "flake8-bugbear==20.1.4"
+          - "flake8==4.0.1"
+          - "flake8-broken-line==0.4.0"
+          - "flake8-bugbear==22.1.11"
           - "flake8-builtins==1.5.3"
-          - "flake8-comprehensions==3.2.3"
-          - "flake8-eradicate==0.4.0"
-          - "flake8-import-order==0.18.1"
+          - "flake8-comprehensions==3.8.0"
+          - "flake8-eradicate==1.2.0"
+          - "flake8-isort==4.1.1"
           - "flake8-mutable==1.2.0"
-          - "flake8-pie==0.5.0"
-          - "flake8-quotes==3.2.0"
+          - "flake8-pie==0.15.0"
+          - "flake8-quotes==3.3.1"
           - "flake8-string-format==0.3.0"
-          - "flake8-tidy-imports==4.1.0"
-          - "flake8-variables-names==0.0.3"
-          - "pep8-naming==0.11.1"
+          - "flake8-tidy-imports==4.6.0"
+          - "flake8-variables-names==0.0.4"
+          - "pep8-naming==0.12.1"
         exclude: ^docs/.*$
 
   - repo: "https://github.com/asottile/yesqa"
-    rev: "v1.2.1"
+    rev: "v1.3.0"
     hooks:
       - id: yesqa
         name: "Lint code (yesqa)"
         additional_dependencies: *flake8_additional_dependencies
         exclude: ^docs/.*$
 
-  - repo: "https://github.com/pre-commit/mirrors-mypy"
-    rev: "v0.782"
+  - repo: "local"
     hooks:
       - id: "mypy"
         name: "Lint code (mypy)"
-        args: ["--python-executable=.venv/bin/python3"]
-        exclude: ^docs/.*$
+        entry: "./.venv/bin/mypy"
+        language: "python"
+        "types": ["python"]
+        require_serial: true
+        exclude: ^(docs|tests)/.*$
