+++ .python-version
@@ -1,1 +1,1 @@
-3.8.5
+3.10.2
