--- mypy.ini
@@ -1,30 +1,0 @@
-[mypy]
-allow_untyped_globals = True
-check_untyped_defs = True
-disallow_any_decorated = True
-disallow_any_generics = True
-disallow_incomplete_defs = True
-disallow_untyped_calls = True
-disallow_untyped_decorators = True
-disallow_untyped_defs = True
-follow_imports = normal
-follow_imports_for_stubs = True
-ignore_missing_imports = True
-namespace_packages = True
-show_column_numbers = True
-show_error_codes = True
-warn_redundant_casts = True
-warn_return_any = True
-warn_unused_ignores = True
-
-[mypy-conftest]
-ignore_errors = True
-
-[mypy-migrations.*]
-ignore_errors = True
-
-[mypy-tests.*]
-ignore_errors = True
-
-[mypy-*.tests.*]
-ignore_errors = True
