+++ pyproject.toml
@@ -1,17 +1,65 @@
 [tool.black]
 line_length = 79
+target_version = ["py37"]
 
 [tool.coverage.run]
 branch = true
 source = ["pre_commit_run_hook_entry"]
 
 [tool.coverage.paths]
-source = ["pre_commit_run_hook_entry.py"]
+source = ["./src/"]
 
 [tool.coverage.report]
 fail_under = 85
+omit = ["./src/pre_commit_run_hook_entry/__main__.py"]
 skip_covered = true
 show_missing = true
+
+[tool.isort]
+profile = "attrs"
+combine_as_imports = true
+force_alphabetical_sort_within_sections = true
+lines_between_types = 0
+src_paths = ["./src/"]
+virtual_env = "./.venv/"
+
+[tool.mypy]
+allow_untyped_globals = true
+check_untyped_defs = true
+disallow_any_decorated = true
+disallow_any_generics = true
+disallow_incomplete_defs = true
+disallow_untyped_calls = true
+disallow_untyped_decorators = true
+disallow_untyped_defs = true
+exclude = "^tests/.*|.*/migrations/.*$"
+follow_imports = "normal"
+follow_imports_for_stubs = true
+ignore_missing_imports = false
+mypy_path = "./src/"
+namespace_packages = true
+python_executable = "./.venv/bin/python3"
+show_column_numbers = true
+show_error_codes = true
+warn_redundant_casts = true
+warn_return_any = true
+warn_unused_ignores = true
+
+[[tool.mypy.overrides]]
+module = [
+  "conftest",
+  "migrations.*",
+  "tests.*",
+  "*.tests",
+  "*.tests.*",
+]
+ignore_errors = true
+
+[[tool.mypy.overrides]]
+module = [
+  "pre_commit.*",
+]
+ignore_missing_imports = true
 
 [tool.poetry]
 name = "pre-commit-run-hook-entry"
@@ -22,7 +70,7 @@ license = "BSD-3-Clause"
 readme = "README.rst"
 homepage = "https://igordavydenko.com/projects/#pre-commit-run-hook-entry"
 repository = "https://github.com/playpauseandstop/pre-commit-run-hook-entry"
-keywords = ["pre-commit", "pre-commit hook", "vscode", "sublime-text3"]
+keywords = ["pre-commit", "pre-commit hook", "vscode", "sublime-text", "sublime-text3"]
 classifiers = [
   "Development Status :: 3 - Alpha",
   "Environment :: Plugins",
@@ -36,25 +84,28 @@ classifiers = [
 
 [tool.poetry.dependencies]
 python = "^3.7"
-pre-commit = "^2.7.1"
+pre-commit = "^2.17.0"
 
 [tool.poetry.dev-dependencies]
-coverage = {extras = ["toml"], version = "^5.3"}
-pytest = "^6.1.2"
-pytest-cov = "^2.9.0"
+coverage = "^6.3.1"
+mypy = "^0.931"
+pytest = "^7.0.1"
+pytest-cov = "^3.0.0"
 
 [tool.poetry.scripts]
-pre-commit-run-black-entry = "pre_commit_run_hook_entry:main_black"
-pre-commit-run-hook-entry = "pre_commit_run_hook_entry:main"
-pre-commit-which-hook-entry = "pre_commit_run_hook_entry:main_which"
+pre-commit-run-black-entry = "pre_commit_run_hook_entry.cli:main_black"
+pre-commit-run-hook-entry = "pre_commit_run_hook_entry.cli:main"
+pre-commit-which-hook-entry = "pre_commit_run_hook_entry.cli:main_which"
 
 [tool.poetry.urls]
 "Bug Tracker" = "https://github.com/playpauseandstop/pre-commit-run-hook-entry/issues"
 
 [tool.pytest.ini_options]
-addopts = "--cov --no-cov-on-fail"
-python_files = "tests.py"
+minversion = "7.0.1"
+testpaths = ["./tests/"]
+addopts = "--cov --cov-config=pyproject.toml --no-cov-on-fail"
+log_level = "info"
 
 [build-system]
-requires = ["poetry>=0.12"]
-build-backend = "poetry.masonry.api"
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
