+++ src/pre_commit_run_hook_entry/__init__.py
@@ -1,0 +1,4 @@
+__prog__ = "pre-commit-run-hook-entry"
+__author__ = "Igor Davdenko"
+__license__ = "BSD-3-Clause"
+__version__ = "1.0.0a3"
