+++ src/pre_commit_run_hook_entry/__main__.py
@@ -1,0 +1,7 @@
+import sys
+
+from pre_commit_run_hook_entry.cli import main
+
+
+if __name__ == "__main__":  # pragma: no cover
+    sys.exit(main())
