+++ src/pre_commit_run_hook_entry/annotations.py
@@ -1,0 +1,4 @@
+from typing import Sequence
+
+
+Argv = Sequence[str]
